Fixes bug in has? of TTL cache with composed cached
If a TTL cache is composed with another cache (with an eviction policy)
the has? function has to check the presence/validity of the key, not
only in the TTL list, but also in the composed cache.

This adds a test with a TTL cache composing a one-item LRU cache. Two
items are added to the composed cache. The TTL has not expired, but the
LRU cache should have removed the first added item. The has? function
should then return false on the first item.